+++ vortex-json/Cargo.toml
@@ -19,6 +19,13 @@ workspace = true
 [dependencies]
 arrow-array = { workspace = true }
 arrow-schema = { workspace = true, features = ["canonical_extension_types"] }
+prost = { workspace = true }
 vortex-array = { workspace = true, default-features = false }
 vortex-error = { workspace = true, default-features = false }
+vortex-proto = { workspace = true, features = ["expr"] }
 vortex-session = { workspace = true }
+
+[dev-dependencies]
+vortex-array = { workspace = true, default-features = false, features = [
+    "_test-harness",
+] }
